+++ test/testOctree.cpp
@@ -40,6 +40,7 @@
  * @return Exit code: 0 on success.
  */
 int main(int argc, char **argv) {
+
   // first test the Box-CoordinateVector distance function
   Box<> box(CoordinateVector<>(), CoordinateVector<>(1.));
   {
@@ -108,10 +109,10 @@ int main(int argc, char **argv) {
     assert_condition(box.periodic_distance(boxCaseH, vecCaseH) == 0.125);
   }
 
-  unsigned int numpos = 100;
+  uint_fast32_t numpos = 100;
   std::vector< CoordinateVector<> > positions(numpos);
   std::vector< double > hs(numpos);
-  for (unsigned int i = 0; i < numpos; ++i) {
+  for (uint_fast32_t i = 0; i < numpos; ++i) {
     positions[i] = Utilities::random_position();
     hs[i] = 0.5 * Utilities::random_double();
   }
@@ -122,22 +123,22 @@ int main(int argc, char **argv) {
     tree.set_auxiliaries(hs, Octree::max< double >);
 
     CoordinateVector<> centre(0.5);
-    std::vector< unsigned int > ngbs_brute_force;
-    for (unsigned int i = 0; i < numpos; ++i) {
+    std::vector< uint_fast32_t > ngbs_brute_force;
+    for (uint_fast32_t i = 0; i < numpos; ++i) {
       double r = (positions[i] - centre).norm();
       if (r < hs[i]) {
         ngbs_brute_force.push_back(i);
       }
     }
-    cmac_status("Number of ngbs (brute force): %lu.", ngbs_brute_force.size());
+    cmac_status("Number of ngbs (brute force): %zd.", ngbs_brute_force.size());
 
-    std::vector< unsigned int > ngbs_tree = tree.get_ngbs(centre);
-    cmac_status("Number of ngbs (tree): %lu.", ngbs_tree.size());
+    std::vector< uint_fast32_t > ngbs_tree = tree.get_ngbs(centre);
+    cmac_status("Number of ngbs (tree): %zd.", ngbs_tree.size());
 
     assert_condition(ngbs_brute_force.size() == ngbs_tree.size());
     std::sort(ngbs_brute_force.begin(), ngbs_brute_force.end());
     std::sort(ngbs_tree.begin(), ngbs_tree.end());
-    for (unsigned int i = 0; i < ngbs_tree.size(); ++i) {
+    for (size_t i = 0; i < ngbs_tree.size(); ++i) {
       assert_condition(ngbs_brute_force[i] == ngbs_tree[i]);
     }
   }
@@ -148,22 +149,22 @@ int main(int argc, char **argv) {
     tree.set_auxiliaries(hs, Octree::max< double >);
 
     CoordinateVector<> centre(0.1);
-    std::vector< unsigned int > ngbs_brute_force;
-    for (unsigned int i = 0; i < numpos; ++i) {
+    std::vector< uint_fast32_t > ngbs_brute_force;
+    for (uint_fast32_t i = 0; i < numpos; ++i) {
       double r = box.periodic_distance(positions[i], centre).norm();
       if (r < hs[i]) {
         ngbs_brute_force.push_back(i);
       }
     }
-    cmac_status("Number of ngbs (brute force): %lu.", ngbs_brute_force.size());
+    cmac_status("Number of ngbs (brute force): %zd.", ngbs_brute_force.size());
 
-    std::vector< unsigned int > ngbs_tree = tree.get_ngbs(centre);
-    cmac_status("Number of ngbs (tree): %lu.", ngbs_tree.size());
+    std::vector< uint_fast32_t > ngbs_tree = tree.get_ngbs(centre);
+    cmac_status("Number of ngbs (tree): %zd.", ngbs_tree.size());
 
     assert_condition(ngbs_brute_force.size() == ngbs_tree.size());
     std::sort(ngbs_brute_force.begin(), ngbs_brute_force.end());
     std::sort(ngbs_tree.begin(), ngbs_tree.end());
-    for (unsigned int i = 0; i < ngbs_tree.size(); ++i) {
+    for (size_t i = 0; i < ngbs_tree.size(); ++i) {
       assert_condition(ngbs_brute_force[i] == ngbs_tree[i]);
     }
   }
